requirements.txt
groq>=0.4.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
